feat: :sparkles: category endpoints

+++ backend/src/main/java/com/tienda/productsale/domain/entity/ProductSalePk.java
@@ -8,8 +8,8 @@ import lombok.Data;
 @Data
 public class ProductSalePk {
   @Column(name = "id_product")
-  private Integer idProduct;
+  private Long idProduct;
 
   @Column(name="id_sale")
-  private Integer idSale;
+  private Long idSale;
 }
