+++ backend/src/main/java/com/tienda/sale/domain/entity/Sale.java
@@ -1,15 +1,5 @@
 package com.tienda.sale.domain.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.PrePersist;
-
 import java.util.Date;
 
 import org.springframework.data.annotation.CreatedDate;
@@ -18,6 +8,14 @@ import org.springframework.data.annotation.LastModifiedDate;
 import com.tienda.customer.domain.entity.Customer;
 
 import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.Data;
 
 @Data
@@ -26,7 +24,7 @@ public class Sale {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name="id_sale")
-  private Integer idSale;
+  private Long idSale;
 
   @ManyToOne
   @JoinColumn(name="id_customer", nullable = false)
@@ -45,4 +43,6 @@ public class Sale {
 
   @Enumerated(EnumType.STRING)
   private SaleStatus status = SaleStatus.PENDING;
+
+  
 }
